requirements.txt
django == 2.2.2
django-environ == 0.4.5
django-taggit == 1.1.0
